requirements.txt
# backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
redis==4.5.4
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.1.0.post1
jinja2==3.1.2
aiofiles==23.2.1
motor==3.3.1
python-dateutil==2.8.2
pytest==7.4.3
httpx==0.25.2
alembic==1.13.0